chore: 型間違え

+++ app/router/graph/model/card.validator.go
@@ -36,7 +36,7 @@ func (i *CardsInput) Validate(ctx context.Context) error {
 	v := validate()
 
 	type where struct {
-		DeckID int `json:"deckId" validate:"required"`
+		DeckID int64 `json:"deckId" validate:"required"`
 	}
 	type input struct {
 		Where  *where `json:"where" validate:"required"`
